+++ scaffold-web/src/main/java/com/xkcoding/web/exception/ServiceException.java
@@ -6,22 +6,21 @@
  *    http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
-package com.xkcoding.log.exception;
+package com.xkcoding.web.exception;
 
 import com.xkcoding.common.api.IResultCode;
 import com.xkcoding.common.api.ResultCode;
 import lombok.Getter;
 
-
 /**
  * <p>
- * 业务异常
+ * 通用业务异常
  * </p>
  *
- * @package: com.xkcoding.log.exception
- * @description: 业务异常
+ * @package: com.xkcoding.web.exception
+ * @description: 通用业务异常
  * @author: yangkai.shen
- * @date: Created in 2019-03-08 14:53
+ * @date: Created in 2019-03-13 13:49
  * @copyright: Copyright (c) 2019
  * @version: V1.0
  * @modified: yangkai.shen
@@ -31,20 +30,31 @@ public class ServiceException extends RuntimeException {
 
     @Getter
     private final IResultCode resultCode;
+    @Getter
+    private final Object data;
 
     public ServiceException(String message) {
         super(message);
         this.resultCode = ResultCode.INTERNAL_SERVER_ERROR;
+        data = null;
     }
 
     public ServiceException(IResultCode resultCode) {
         super(resultCode.getMessage());
         this.resultCode = resultCode;
+        data = null;
+    }
+
+    public ServiceException(IResultCode resultCode, Object data) {
+        super(resultCode.getMessage());
+        this.resultCode = resultCode;
+        this.data = data;
     }
 
     public ServiceException(IResultCode resultCode, Throwable cause) {
         super(cause);
         this.resultCode = resultCode;
+        data = null;
     }
 
     /**